requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
django==5.0.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r requirements.in
djangorestframework==3.15.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
